moved tool bar
