+++ projects/aca-shared/src/lib/components/tool-bar/toolbar-menu-item/toolbar-menu-item.component.ts
@@ -25,30 +25,26 @@
 
 import { Component, Input, ViewEncapsulation } from '@angular/core';
 import { ContentActionRef } from '@alfresco/adf-extensions';
-import { AppExtensionService } from '../../../extensions/extension.service';
-
-export enum ToolbarButtonType {
-  ICON_BUTTON = 'icon-button',
-  MENU_ITEM = 'menu-item'
-}
+import { SharedExtensionService } from '../../../services/shared-extension.service';
 
 @Component({
-  selector: 'app-toolbar-button',
-  templateUrl: 'toolbar-button.component.html',
+  selector: 'app-toolbar-menu-item',
+  templateUrl: 'toolbar-menu-item.component.html',
+  styles: [
+    `
+      .app-toolbar-menu-item:last-child > .mat-divider-horizontal {
+        display: none;
+      }
+    `
+  ],
   encapsulation: ViewEncapsulation.None,
-  host: { class: 'app-toolbar-button' }
+  host: { class: 'app-toolbar-menu-item' }
 })
-export class ToolbarButtonComponent {
-  @Input()
-  type: ToolbarButtonType = ToolbarButtonType.ICON_BUTTON;
-
-  @Input()
-  color = '';
-
+export class ToolbarMenuItemComponent {
   @Input()
   actionRef: ContentActionRef;
 
-  constructor(private extensions: AppExtensionService) {}
+  constructor(private extensions: SharedExtensionService) {}
 
   runAction() {
     if (this.hasClickAction(this.actionRef)) {
@@ -61,5 +57,9 @@ export class ToolbarButtonComponent {
       return true;
     }
     return false;
+  }
+
+  trackById(_: number, obj: { id: string }) {
+    return obj.id;
   }
 }
